+++ dbusutil/gsprop/gsprop.go
@@ -8,7 +8,7 @@ import (
 	"github.com/godbus/dbus"
 	"github.com/linuxdeepin/go-lib/dbusutil"
 	"github.com/linuxdeepin/go-lib/gsettings"
-	"pkg.deepin.io/gir/gio-2.0"
+	"github.com/linuxdeepin/go-gir/gio-2.0"
 )
 
 type base struct {