+++ dbusutil/gsprop/gsprop_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/godbus/dbus"
 	"github.com/godbus/dbus/introspect"
-	"pkg.deepin.io/gir/gio-2.0"
+	"github.com/linuxdeepin/go-gir/gio-2.0"
 	"github.com/linuxdeepin/go-lib/dbusutil"
 	"github.com/linuxdeepin/go-lib/gsettings"
 )