+++ utils/keyfile.go
@@ -24,7 +24,7 @@ import (
 	"os"
 	"sync"
 
-	"pkg.deepin.io/gir/glib-2.0"
+	"github.com/linuxdeepin/go-gir/glib-2.0"
 )
 
 var (